backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
mongoengine
python-dotenv 